<fix> return

+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -5,9 +5,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.Collection;
 
@@ -21,31 +22,50 @@ public class FilmService {
     @Autowired
     private final FilmStorage filmStorage;
 
+    @Autowired
+    private final UserStorage userStorage;
+
     public Film create(Film film) {
         return filmStorage.create(film);
     }
 
     public Film update(Film film) {
-        return filmStorage.update(film);
+        if (filmStorage.getById(film.getId()) != null) {
+            return filmStorage.update(film);
+        } else {
+            throw new NotFoundException("Данные не найдены");
+        }
     }
 
-    public void delete(long id){
-        filmStorage.delete(id);
+    public void delete(Long id) {
+        if (filmStorage.getById(id) != null) {
+            filmStorage.delete(id);
+        } else {
+            throw new NotFoundException("Данные не найдены");
+        }
     }
 
     public Collection<Film> getAll() {
         return filmStorage.getFilms();
     }
 
-    public void addLike(long id, long userId) {
-        filmStorage.addLike(id,userId);
+    public Film addLike(Long id, Long userId) {
+        if (userStorage.get(userId) != null && filmStorage.getById(id) != null) {
+            return filmStorage.addLike(id, userId);
+        } else {
+            throw new NotFoundException("Данные не найдены");
+        }
     }
 
-    public void deleteLike(long id, long userId){
-        filmStorage.deleteLike(id,userId);
+    public Film deleteLike(Long id, Long userId) {
+        if (userStorage.get(userId) != null && filmStorage.getById(id) != null) {
+            return filmStorage.deleteLike(id, userId);
+        } else {
+            throw new NotFoundException("Данные не найдены");
+        }
     }
 
-    public Collection<Film> getPopularFilm(Integer count){
+    public Collection<Film> getPopularFilm(int count) {
         return filmStorage.getPorularFilm(count);
     }
 }
